% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophylogenetic_signal.R
\name{gather_parafit_stats}
\alias{gather_parafit_stats}
\title{Generate summary stats (pval means, etc.) for the results of run_parafit_loop()
Based on https://github.com/alixmatthews/cophylogenetic/blob/master/parafit_script.R}
\usage{
gather_parafit_stats(res, assoc_mat)
}
\arguments{
\item{res}{results from run_parafit_loop}

\item{assoc_mat}{association matrix, with hosts as rownames and symbionts as colnames}
}
\description{
Generate summary stats (pval means, etc.) for the results of run_parafit_loop()
Based on https://github.com/alixmatthews/cophylogenetic/blob/master/parafit_script.R
}
