% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distances.R
\name{get_cophenetic}
\alias{get_cophenetic}
\title{Calculate cophenetic distances for a phylogenetic tree}
\usage{
get_cophenetic(phylo, taxa_sort)
}
\arguments{
\item{phylo}{object of class phylo}

\item{taxa_sort}{character vector of the taxa, in order, for which to get distances.}
}
\description{
This function gets the cophenetic distances of a phylo class object.
}
