% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophylogenetic_signal.R
\name{get_pcoas}
\alias{get_pcoas}
\title{PACO functions
Transforms the host and parasite distance matrices into the respective matrices of Principal Coordinates (ape::pcoa)
and duplicates taxa (if necessary) to accommodate multiple host-parasite associations}
\usage{
get_pcoas(host_dist, symbiont_dist, assoc_mat)
}
\description{
PACO functions
Transforms the host and parasite distance matrices into the respective matrices of Principal Coordinates (ape::pcoa)
and duplicates taxa (if necessary) to accommodate multiple host-parasite associations
}
