% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_data.R
\name{read_associations}
\alias{read_associations}
\title{Read in an association matrix}
\usage{
read_associations(filepath)
}
\arguments{
\item{filepath}{path to association matrix (CSV format) with hosts as rownames and symbionts as colnames.}
}
\description{
Read in an association matrix
}
