% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_data.R
\name{read_phylo}
\alias{read_phylo}
\title{Read in phylogenetic tree from a .nexus file}
\usage{
read_phylo(filepath, nexus_format, taxa_to_keep = NULL, label_fix = NULL)
}
\arguments{
\item{filepath}{path to the nexus file}

\item{nexus_format}{exus format to be pass to \code{phytools::readNexus()}, either 'standard' or 'raxml'}

\item{taxa_to_keep}{(optional) a list (vector) of taxa in the tree to keep, if left blank, all taxa will be kept}

\item{label_fix}{(optional) a character vector (length = 2) of strings that can be used to replace string prefixes/suffixes in the data; e.g., c(' ', '_') to replace spaces with underscores}
}
\description{
Read in phylogenetic tree from a .nexus file
}
