% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophylogenetic_signal.R
\name{run_parafit_loop}
\alias{run_parafit_loop}
\title{Multiple parafit runs.
Use lapply to run the single_parafit_run() multiple times (n_runs)
Based on https://github.com/alixmatthews/cophylogenetic/blob/master/parafit_script.R}
\usage{
run_parafit_loop(
  host_dist,
  symbiont_dist,
  assoc_mat,
  n_runs,
  correction_method
)
}
\arguments{
\item{host_dist}{host distance matrix}

\item{symbiont_dist}{symbiont distance matrix}

\item{assoc_mat}{association matrix, with hosts as rownames and symbionts as colnames}
}
\value{
a list of results from individual parafit runs (`single_parafit_run()``)
}
\description{
Multiple parafit runs.
Use lapply to run the single_parafit_run() multiple times (n_runs)
Based on https://github.com/alixmatthews/cophylogenetic/blob/master/parafit_script.R
}
