% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophylogenetic_signal.R
\name{single_parafit_run}
\alias{single_parafit_run}
\title{A single run of the parafit function
Based on https://github.com/alixmatthews/cophylogenetic/blob/master/parafit_script.R}
\usage{
single_parafit_run(i, host_dist, symbiont_dist, assoc_mat, correction_method)
}
\description{
A single run of the parafit function
Based on https://github.com/alixmatthews/cophylogenetic/blob/master/parafit_script.R
}
